utilise slideType plutot que calendarSliceType
