+++ frontend/app/components/ui/commons/selectBar/selectBar.vue
@@ -4,7 +4,6 @@ import YearPicker from "./yearPicker.vue";
 import StackedYearPicker from "./stackedYearPicker.vue";
 import DayPicker from "./dayPicker.vue";
 import SliceType from "./sliceType.vue";
-import CalendarSliceType from "./calendarSliceType.vue";
 import RecordsPeriodSlice from "./recordsPeriodSlice.vue";
 import ExportMenu from "~/components/ui/commons/exportMenu.vue";
 import type {
@@ -219,7 +218,7 @@ const granularityValues = computed(() => [
                                 }}
                             </template>
                         </USwitch>
-                        <CalendarSliceType
+                        <SliceType
                             v-if="adapter.calendarAverageEnabled?.value"
                         />
                     </div>
